requirements.txt
numpy>=1.22.4
pandas>=1.1.5
pytest==7.1.3
zenml==0.20.4
Pillow==9.2.0
tensorflow>=2.0.0
PyYAML~=5.4.1
mlflow~=1.25.1